Shotgun/requirements.txt
pytorch
einops